+++ engine/arch/simd/simd.cpp
@@ -1,4 +1,4 @@
-#include "../arch.hpp"
+#include "simd.hpp"
 
 #if defined(TARGET_ARM_NEON)
     #include "neon.hpp"
@@ -7,5 +7,5 @@
 #elif defined(TARGET_X86_AVX2)
     #include "avx2.hpp"
 #else
-    #include "fallback.hpp"
+    #include "scalar.hpp"
 #endif
